In packets from inputnode to correlator node the packet sizes were encoded as type size_t but this fails when mixing 32bit and 64bit nodes.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@977 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/input_node_data_writer.h
@@ -102,7 +102,7 @@ private:
   int bytes_to_write(int byte_offset, int next_delay_pos);
   int get_next_delay_pos(std::vector<Delay> &cur_delay, uint64_t start_time);
 
-  void write_header(Data_writer_sptr writer, size_t ndata, int inv_start, int nr_inv, int delay);
+  void write_header(Data_writer_sptr writer, int32_t ndata, int inv_start, int nr_inv, int delay);
   void write_random_data(Data_writer_sptr writer, int ndata);
   void write_data(Data_writer_sptr writer, int ndata, int byte_offset);
   void write_invalid_blocks(Data_writer_sptr writer, int byte_offset, int n_bytes,
